+++ src/App.tsx
@@ -1,12 +1,15 @@
 import AppRouter from './app/router'
 import OfflineModal from './app/OfflineModal'
 import SessionProvider from './features/auth/SessionProvider'
+import ToastProvider from './shared/toast/ToastProvider'
 
 function App() {
   return (
     <SessionProvider>
-      <AppRouter />
-      <OfflineModal />
+      <ToastProvider>
+        <AppRouter />
+        <OfflineModal />
+      </ToastProvider>
     </SessionProvider>
   )
 }